+++ src/types/Star.type.ts
@@ -1,0 +1,5 @@
+export type Star = {
+  number: number;
+  title: string;
+  handleRatingChange: (number: number) => void;
+};
